solution/requirements.txt
spacy
python-docx
pandas